refactor(reservation-domain): Seat 패키지로 도메인 분리
- Seat, SeatTimeSlot을 대상으로 domain.seat 패키지로 이동함
- Test, TestFixture 도 패키지 맞춤
