+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/availability/model/AvailableSeats.java
@@ -1,7 +1,7 @@
 package com.seatliberator.seatliberator.reservation.application.availability.model;
 
 import com.seatliberator.seatliberator.reservation.domain.SeatLocator;
-import com.seatliberator.seatliberator.reservation.domain.persistence.Seat;
+import com.seatliberator.seatliberator.reservation.domain.seat.Seat;
 
 import java.util.List;
 import java.util.stream.Collectors;
