+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/book/application/ReservationCommandServiceTest.java
@@ -19,8 +19,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Optional;
 
-import static com.seatliberator.seatliberator.reservation.domain.fixture.SeatFixture.createSeat;
 import static com.seatliberator.seatliberator.reservation.domain.fixture.TestSupport.fixedClock;
+import static com.seatliberator.seatliberator.reservation.domain.seat.SeatFixture.createSeat;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
