+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/room/application/SeatAssignmentServiceTest.java
@@ -7,9 +7,9 @@ import com.seatliberator.seatliberator.reservation.application.room.port.out.Sea
 import com.seatliberator.seatliberator.reservation.application.shared.exception.ReservationApplicationErrorCode;
 import com.seatliberator.seatliberator.reservation.application.shared.exception.ReservationApplicationException;
 import com.seatliberator.seatliberator.reservation.domain.fixture.RoomFixture;
-import com.seatliberator.seatliberator.reservation.domain.fixture.SeatFixture;
 import com.seatliberator.seatliberator.reservation.domain.persistence.Room;
-import com.seatliberator.seatliberator.reservation.domain.persistence.Seat;
+import com.seatliberator.seatliberator.reservation.domain.seat.Seat;
+import com.seatliberator.seatliberator.reservation.domain.seat.SeatFixture;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
