+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/room/application/SeatQueryServiceTest.java
@@ -23,8 +23,8 @@ import java.util.List;
 import java.util.Optional;
 
 import static com.seatliberator.seatliberator.reservation.domain.fixture.RoomFixture.createRoom;
-import static com.seatliberator.seatliberator.reservation.domain.fixture.SeatFixture.create;
 import static com.seatliberator.seatliberator.reservation.domain.fixture.TestSupport.fixedClock;
+import static com.seatliberator.seatliberator.reservation.domain.seat.SeatFixture.create;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
