--- reservation/reservation-domain/src/main/java/com/seatliberator/seatliberator/reservation/domain/SeatStatus.java
@@ -1,6 +1,0 @@
-package com.seatliberator.seatliberator.reservation.domain;
-
-public enum SeatStatus {
-    ACTIVE,
-    INACTIVE
-}
