+++ reservation/reservation-domain/src/main/java/com/seatliberator/seatliberator/reservation/domain/seat/SeatTimeSlot.java
@@ -1,4 +1,4 @@
-package com.seatliberator.seatliberator.reservation.domain.persistence;
+package com.seatliberator.seatliberator.reservation.domain.seat;
 
 import com.seatliberator.seatliberator.kernel.condition.Preconditions;
 import com.seatliberator.seatliberator.reservation.domain.EmbeddableTimeRange;
