+++ reservation/reservation-domain/src/main/java/com/seatliberator/seatliberator/reservation/domain/seat/SeatTimeSlotStatus.java
@@ -1,0 +1,6 @@
+package com.seatliberator.seatliberator.reservation.domain.seat;
+
+public enum SeatTimeSlotStatus {
+    ACTIVE,
+    INACTIVE
+}
