+++ reservation/reservation-domain/src/test/java/com/seatliberator/seatliberator/reservation/domain/persistence/RoomTest.java
@@ -15,7 +15,7 @@ import java.util.stream.Stream;
 
 import static com.seatliberator.seatliberator.kernel.test.assertion.DomainAssertions.assertThatDomainThrownBy;
 import static com.seatliberator.seatliberator.reservation.domain.fixture.RoomFixture.INITIAL_ROOM_ID;
-import static com.seatliberator.seatliberator.reservation.domain.fixture.SeatFixture.INITIAL_CREATED_AT;
+import static com.seatliberator.seatliberator.reservation.domain.seat.SeatFixture.INITIAL_CREATED_AT;
 import static com.seatliberator.seatliberator.reservation.domain.fixture.TestSupport.fixedClock;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
